+++ environments/gloo-gateway/gateway-api/1.19/with-istio-helm-ambient/gloo-enterprise-app/init.sh
@@ -46,4 +46,4 @@ EOF
 
 # Install Gateway API CRDs
 echo "installing Gateway API CRDs"
-kubectl apply -f https://github.com/kubernetes-sigs/gateway-api/releases/download/v1.2.0/standard-install.yaml
+kubectl apply -f https://github.com/kubernetes-sigs/gateway-api/releases/download/v1.2.1/standard-install.yaml