+++ environments/gloo-gateway/gateway-api/1.19/with-istio-helm-ambient/istio/init.sh
@@ -6,4 +6,4 @@ echo "istiod and ingress gateway deployments"
 # install Kubernetes Gateway CRDs
 echo "installing Kubernetes Gateway CRDs"
 kubectl get crd gateways.gateway.networking.k8s.io &> /dev/null || \
-  { kubectl apply -f https://github.com/kubernetes-sigs/gateway-api/releases/download/v1.2.0/standard-install.yaml; }
+  { kubectl apply -f https://github.com/kubernetes-sigs/gateway-api/releases/download/v1.2.1/standard-install.yaml; }