+++ environments/istio/ambient-demo/core/istio/gke/init.sh
@@ -9,7 +9,7 @@ kubectl create ns istio-system
 # install Kubernetes Gateway CRDs
 echo "installing Kubernetes Gateway CRDs"
 kubectl get crd gateways.gateway.networking.k8s.io &> /dev/null || \
-  { kubectl apply -f https://github.com/kubernetes-sigs/gateway-api/releases/download/v1.2.0/standard-install.yaml; }
+  { kubectl apply -f https://github.com/kubernetes-sigs/gateway-api/releases/download/v1.2.1/standard-install.yaml; }
 
 # apply resourcequota for istio-system
 kubectl apply -f - <<EOF