+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - update /gloo-mesh-core environments to use gloo-platform 2.7.3
 - update /istio environments to use istio 1.26.0
 - update /gloo-gateway environments to use gloo-gateway 1.19.0
+- update gloo-operator to 0.2.3 in gateway-api/1.19/with-ilm-ambient environment
+- update gwapi crds to use v1.2.1 in all environments
 
 0.8.7 (4-22-25)
 ---